requirements.txt
openai 
langchain 
faiss-cpu 
sentence-transformers 
sympy 
langchain_openai 
langchain-community
pytesseract 
pillow
typhoon-ocr
fastapi
python-dotenv